producer/requirements.txt
kafka-python
pyyaml
confluent_kafka
schedule
pandas
scipy
sqlalchemy